fix(claude-monitor): fix platformio compile — remove USER_SETUP_LOADED, git XPT2046, fix narrowing warning
- Remove -DUSER_SETUP_LOADED build flag (caused TFT_eSPI to skip User_Setup.h)
- Change XPT2046_Touchscreen to git URL (registry version too old, lacks begin(SPIClass&))
- Cast ternary to uint16_t in budget.cpp to fix narrowing conversion warning

+++ apps/claude-monitor/src/screens/budget.cpp
@@ -61,7 +61,7 @@ void screenBudgetDraw(const UsageData& data) {
   int gx[] = {4, 110, 216};
   const char* labels[] = {"SISA", "RATA/HARI", "PROYEKSI"};
   float values[] = {sisa, rataHarian, proyeksi};
-  uint16_t vColors[] = {C_GREEN, C_YELLOW, (proyeksi > budget ? C_RED : C_YELLOW)};
+  uint16_t vColors[] = {C_GREEN, C_YELLOW, (uint16_t)(proyeksi > budget ? C_RED : C_YELLOW)};
 
   for (int i = 0; i < 3; i++) {
     tft.fillRect(gx[i], 124, 102, 52, bgGrid);
